Refactor source folder structure
