+++ src/management/JobsManager.js
@@ -3,7 +3,7 @@ var extend = require('util')._extend;
 var Promise = require('bluebird');
 
 var RestClient = require('rest-facade').Client;
-var ArgumentError = require('./exceptions').ArgumentError;
+var ArgumentError = require('../exceptions').ArgumentError;
 
 
 /**
